refactor(auth): 移除已废弃的权限变更事件发布器相关代码
- 删除 PermissionChangePublisher Bean 及相关注释
- 修改 FlowDefinitionCacheBroadcaster 中注释，替换权限变更事件发布器引用为 PermissionChangeNotifier
- 更新 PermissionChangeCacheInvalidator 注释中的引用为 PermissionChangeNotifier

perf(safe): 集成 Sentinel 限流扩展和补充安全模块指标监控

- 在安全限流模块添加 Sentinel 限流器实现 SentinelRateLimiter
- pom.xml 中引入 sentinel-core 依赖，支持高级分布式限流
- 在安全模块健康检查 SafeHealthIndicator 中添加安全指标累计值展示
- 新增安全模块指标文档 SafeMetricsDoc，定义安全 SLI、SLO 和告警规则

refactor(safe-core): 删除安全事件环形缓冲区相关代码

- 移除 SecurityEventRingBuffer 及其对 SecurityEventPublisher 的引用和相关方法
- 简化 SecurityEventPublisher，去除环形缓冲区快照和发布耗时统计功能

chore(seata): 更新依赖和日志变量命名

- pom.xml 中添加 rocketmq-client 可选依赖支持 MqXidPropagator
- 修改 SeataAutoConfiguration 中日志变量名为 LOG
- 修正 SeataHealthIndicator 中 pendingCount 类型为 long

docs(safe): 补充安全模块 README 中模块结构说明

- 新增安全模块各子模块介绍和使用方式说明，提升文档完整性

refactor(audit): 删除已废弃的 AuditStorage 接口及相关注释

- 完全移除 AuditStorage.java 文件，推荐使用 AuditWriter 替代

+++ ydsz-common/ydsz-common-seata/src/main/java/com/njydsz/common/seata/health/SeataHealthIndicator.java
@@ -112,7 +112,7 @@ public class SeataHealthIndicator implements HealthIndicator {
         }
         LocalDateTime threshold = LocalDateTime.now().minusNanos(
                 properties.getRecoveryTimeoutThresholdMs() * 1_000_000);
-        int pendingCount = logStore.countTimeoutPending(threshold);
+        long pendingCount = logStore.countTimeoutPending(threshold);
         builder.withDetail("tccPendingTransactions", pendingCount);
         builder.withDetail("tccRecoveryScanIntervalMs", properties.getRecoveryScanIntervalMs());
